JBAI-13625. Extract Closeable interface into a separate shared module
