+++ agents/agents-core/src/commonMain/kotlin/ai/grazie/code/agents/core/agent/AIAgent.kt
@@ -8,22 +8,25 @@ import ai.grazie.code.agents.core.environment.AIAgentEnvironmentUtils.mapToToolR
 import ai.grazie.code.agents.core.environment.ReceivedToolResult
 import ai.grazie.code.agents.core.environment.TerminationTool
 import ai.grazie.code.agents.core.exception.AgentEngineException
-import ai.grazie.code.agents.core.feature.AIAgentPipeline
 import ai.grazie.code.agents.core.feature.AIAgentFeature
+import ai.grazie.code.agents.core.feature.AIAgentPipeline
 import ai.grazie.code.agents.core.model.AgentServiceError
 import ai.grazie.code.agents.core.model.AgentServiceErrorType
 import ai.grazie.code.agents.core.model.message.*
 import ai.grazie.code.agents.core.tools.*
 import ai.grazie.code.agents.core.tools.annotations.InternalAgentToolsApi
 import ai.grazie.code.agents.core.tools.tools.ToolStage
-import ai.grazie.code.agents.core.utils.Closeable
 import ai.grazie.code.agents.local.features.common.config.FeatureConfig
+import ai.grazie.code.agents.utils.ai.grazie.code.agents.utils.Closeable
 import ai.grazie.utils.mpp.LoggerFactory
 import ai.grazie.utils.mpp.UUID
 import ai.jetbrains.code.prompt.executor.model.PromptExecutor
 import ai.jetbrains.code.prompt.message.Message
 import ai.jetbrains.code.prompt.text.TextContentBuilder
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CompletableDeferred
+import kotlinx.coroutines.async
+import kotlinx.coroutines.awaitAll
+import kotlinx.coroutines.supervisorScope
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlinx.serialization.json.JsonObject
