+++ agents/agents-features/agents-features-common/src/commonMain/kotlin/ai/grazie/code/agents/local/features/common/message/FeatureMessageProcessor.kt
@@ -1,5 +1,7 @@
 package ai.grazie.code.agents.local.features.common.message
 
+import ai.grazie.code.agents.utils.ai.grazie.code.agents.utils.Closeable
+
 /**
  * Represents a provider responsible for handling outbound feature messages or events.
  *
