+++ agents/agents-features/agents-features-common/src/jvmTest/kotlin/ai/grazie/code/agents/local/features/remote/client/FeatureMessageRemoteClientTest.kt
@@ -3,12 +3,12 @@ package ai.grazie.code.agents.local.features.remote.client
 import ai.grazie.code.agents.local.features.NetUtil.findAvailablePort
 import ai.grazie.code.agents.local.features.common.message.FeatureMessage
 import ai.grazie.code.agents.local.features.common.message.FeatureStringMessage
-import ai.grazie.code.agents.local.features.common.message.use
 import ai.grazie.code.agents.local.features.common.remote.client.FeatureMessageRemoteClient
 import ai.grazie.code.agents.local.features.common.remote.client.config.DefaultClientConnectionConfig
 import ai.grazie.code.agents.local.features.common.remote.server.FeatureMessageRemoteServer
 import ai.grazie.code.agents.local.features.common.remote.server.config.DefaultServerConnectionConfig
 import ai.grazie.code.agents.local.features.writer.TestFeatureEventMessage
+import ai.grazie.code.agents.utils.ai.grazie.code.agents.utils.use
 import ai.grazie.utils.mpp.LoggerFactory
 import ai.grazie.utils.mpp.MPPLogger
 import io.ktor.http.*
