+++ agents/agents-features/agents-features-common/src/jvmTest/kotlin/ai/grazie/code/agents/local/features/writer/FeatureMessageLogWriterTest.kt
@@ -3,9 +3,9 @@ package ai.grazie.code.agents.local.features.writer
 import ai.grazie.code.agents.local.features.common.message.FeatureEvent
 import ai.grazie.code.agents.local.features.common.message.FeatureMessage
 import ai.grazie.code.agents.local.features.common.message.FeatureStringMessage
-import ai.grazie.code.agents.local.features.common.message.use
 import ai.grazie.code.agents.local.features.common.writer.FeatureMessageLogWriter
 import ai.grazie.code.agents.local.features.common.writer.FeatureMessageLogWriter.LogLevel
+import ai.grazie.code.agents.utils.ai.grazie.code.agents.utils.use
 import ai.grazie.utils.mpp.MPPLogger
 import kotlinx.coroutines.runBlocking
 import kotlin.test.AfterTest
