+++ agents/agents-features/agents-features-trace/src/jvmTest/kotlin/ai/grazie/code/agents/local/features/tracing/writer/TraceFeatureMessageFileWriterTest.kt
@@ -7,8 +7,8 @@ import ai.grazie.code.agents.core.feature.model.*
 import ai.grazie.code.agents.local.features.common.message.FeatureEvent
 import ai.grazie.code.agents.local.features.common.message.FeatureMessage
 import ai.grazie.code.agents.local.features.common.message.FeatureStringMessage
-import ai.grazie.code.agents.local.features.common.message.use
 import ai.grazie.code.agents.local.features.tracing.feature.Tracing
+import ai.grazie.code.agents.utils.ai.grazie.code.agents.utils.use
 import ai.grazie.code.files.jvm.JVMFileSystemProvider
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.io.TempDir
