+++ agents/agents-utils/build.gradle.kts
@@ -1,0 +1,35 @@
+import ai.grazie.gradle.publish.maven.Publishing.publishToGraziePublicMaven
+
+group = "${rootProject.group}.agents"
+version = rootProject.version
+
+plugins {
+    id("ai.kotlin.multiplatform")
+    alias(libs.plugins.kotlin.serialization)
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+        }
+
+        jvmMain {
+        }
+
+        commonTest {
+            dependencies {
+                implementation(kotlin("test"))
+                implementation(libs.kotlinx.coroutines.test)
+            }
+        }
+
+        jvmTest {
+            dependencies {
+                implementation(kotlin("test-junit5"))
+            }
+        }
+    }
+
+}
+
+publishToGraziePublicMaven()
